+++ src/pages/PageCallManage.tsx
@@ -33,6 +33,7 @@ import {
 import { BrekekeGradient } from '../components/BrekekeGradient'
 import { ButtonIcon } from '../components/ButtonIcon'
 import { IncomingItemWithTimer } from '../components/CallNotify'
+import { CallVideosCarousel } from '../components/CallVideosCarousel'
 import { FieldButton } from '../components/FieldButton'
 import { Layout } from '../components/Layout'
 import { RnTouchableOpacity } from '../components/Rn'
@@ -46,6 +47,7 @@ import { getCallStore } from '../stores/callStore'
 import { intl } from '../stores/intl'
 import { Nav } from '../stores/Nav'
 import { Duration } from '../stores/timerStore'
+import { checkMutedRemoteUser } from '../utils/checkMutedRemoteUser'
 import { BrekekeUtils } from '../utils/RnNativeModules'
 import { waitTimeout } from '../utils/waitTimeout'
 import { PageCallTransferAttend } from './PageCallTransferAttend'
@@ -253,6 +255,11 @@ class PageCallManage extends Component<{
   call: Call
 }> {
   componentDidMount = () => {
+    // handle the case when app is killed and opened during a call with incoming call
+    if (this.props.call.incoming) {
+      this.onAppStateChange(AppState.currentState)
+    }
+
     this.checkJavaPn()
     this.componentDidUpdate()
     this.appStateSubscription = AppState.addEventListener(
@@ -344,8 +351,8 @@ class PageCallManage extends Component<{
   }
 
   private appStateSubscription?: NativeEventSubscription
-  private onAppStateChange = () => {
-    if (AppState.currentState === 'active') {
+  private onAppStateChange = (nextAppState: string) => {
+    if (nextAppState === 'active') {
       const { call: c } = this.props
       if (
         this.hasJavaPn &&
@@ -433,6 +440,7 @@ class PageCallManage extends Component<{
         {c.localVideoEnabled && this.renderVideo()}
         {this.renderAvatar()}
         {this.renderBtns()}
+        {c.localVideoEnabled && <View style={{ flex: 1 }} />}
         {this.renderHangupBtn()}
         {c.transferring ? renderTransferring() : null}
       </>
@@ -455,11 +463,27 @@ class PageCallManage extends Component<{
           />
         </View>
         <View style={css.Video_Space} />
-        <View style={css.Video}>
-          <VideoPlayer sourceObject={c.remoteVideoStreamObject} />
+        <View style={[css.Video]}>
+          <VideoPlayer
+            sourceObject={
+              checkMutedRemoteUser(
+                c.remoteUserOptionsTable?.[c.videoStreamActive?.user ?? '']
+                  ?.muted,
+              )
+                ? c.videoStreamActive?.remoteStreamObject
+                : null
+            }
+            zOrder={0}
+          />
         </View>
+        <CallVideosCarousel
+          call={c}
+          showButtonsInVideoCall={this.showButtonsInVideoCall}
+          onButtonsInVideo={this.toggleButtons}
+        />
         <RnTouchableOpacity
           onPress={this.toggleButtons}
+          activeOpacity={0}
           style={[StyleSheet.absoluteFill, { zIndex: 10 }]}
         />
       </>
@@ -477,7 +501,7 @@ class PageCallManage extends Component<{
       : { flex: 1 }
     const styleViewAvatar = isLarge ? styleBigAvatar : css.smallAvatar
     return (
-      <View style={[css.Image_wrapper, { flex: 1 }]}>
+      <View style={[!c.localVideoEnabled && css.Image_wrapper, { flex: 1 }]}>
         <View
           style={isShowAvatar ? styleViewAvatar : { height: 0, opacity: 0 }}
         >
@@ -587,7 +611,9 @@ class PageCallManage extends Component<{
               name={intl`VIDEO`}
               noborder
               onPress={c.toggleVideo}
-              path={c.localVideoEnabled ? mdiVideo : mdiVideoOff}
+              path={
+                c.localVideoEnabled && !c.mutedVideo ? mdiVideo : mdiVideoOff
+              }
               size={40}
               textcolor='white'
             />
@@ -680,7 +706,9 @@ class PageCallManage extends Component<{
     const { call: c } = this.props
     const incoming = c.incoming && !c.answered
     const isLarge = !!(c.partyImageSize && c.partyImageSize === 'large')
-    const isHangupBtnHidden = incoming && this.isBtnHidden('hangup')
+    const isHangupBtnHidden =
+      (incoming && this.isBtnHidden('hangup')) ||
+      (!this.showButtonsInVideoCall && c.answered)
     return (
       <View style={[css.viewHangupBtns, { marginTop: isLarge ? 10 : 40 }]}>
         {c.holding && !c.rqLoadings['hold'] ? (