+++ src/utils/callkeep.ts
@@ -163,6 +163,12 @@ export const setupCallKeepEvents = async () => {
   ) => {
     const uuid = e.callUUID.toUpperCase()
     const c = cs.calls.find(_ => _.callkeepUuid === uuid)
+    // in case the user has not answered the call on callkeep display incoming call
+    // audio session should not be assigned to WebRTC
+    // before the didActivateAudioSession event is called
+    if (Platform.OS === 'ios' && c?.isAutoAnswer && !c.isAudioActive) {
+      return
+    }
     if (c && c.holding !== e.hold) {
       c.toggleHoldWithCheck()
     }
@@ -185,6 +191,11 @@ export const setupCallKeepEvents = async () => {
   ) => {
     // only in ios
     console.log('CallKeep debug: didActivateAudioSession')
+    const c = cs.getOngoingCall()
+    if (c?.isAutoAnswer) {
+      c.isAudioActive = true
+      c.partyAnswered && c.setHoldWithoutCallKeep(false)
+    }
     BrekekeUtils.webrtcSetAudioEnabled(true)
   }
   const didDeactivateAudioSession = (
@@ -348,6 +359,9 @@ export const setupCallKeepEvents = async () => {
       RnStacker.stacks = [RnStacker.stacks[0]]
     }
     cs.inPageCallManage = undefined
+  })
+  eventEmitter.addListener('updateStreamActive', (vId: string) => {
+    cs.getOngoingCall()?.updateVideoStreamFromNative(vId)
   })
   eventEmitter.addListener('debug', (m: string) =>
     console.log(`Android debug: ${m}`),