+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+#### 2.16.3
+
+- Improve debug log with multiples files, and mark for login logout
+- Fix it should login into the correct account when press on the missed call notification (issue 955)
+- Fix it should update avatar without cache in the next call (issue 958)
+- Fix it should display name from phone book in calls (issue 960)
+- Fix android it should have notification sound after a missed group call answer simultaneously (issue 974)
+- Fix ios auto answer it should have audio (issue 975, 978, 1008)
+- Fix it should show UC notification in call manage screen (issue 982)
+- Fix it should remove PN token if login from another device even if there is an ongoing call (issue 1003)
+- Fix android it should resolve dialer permission automatically when there is an incoming call (issue 1006)
+- Fix android it should open app when press on the running in background notification (issue 1022, 1034)
+
+#### 2.16.2
+
+- Embed:
+  - Add support for user agent in embed login options
+
+#### 2.16.1
+
+- Upgrade webrtcclient to 2.0.37.356
+- Fix video conference it should work with UC desktop
+- Fix web browser access front camera on mobile devices
+
+#### 2.16.0
+
+- Initial implementation of video conference
+
 #### 2.15.10
 
 - Fix ios it should not crash when press hold